requirements.txt
matplotlib==3.9.2
numpy==2.1.3
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
segformer_pytorch==0.0.6
torch==2.4.1
torchmetrics==1.5.1
torchvision==0.19.1
tqdm==4.64.1
transformers==4.44.2
